requirements.txt
Django==4.2
djangorestframework==3.14.0
translate==3.6.1
